add basic report list by day, hour, or minute
This is gonna need to be paginated, as there can be thousands of reports
a day. In that case, we're also gonna need some better way of handling
memory: Loading all of the reports so they can be written out in bulk at
once is less efficient than if we can read and write one at a time.